refactor: simplify tooling and type layout
