+++ biome.json
@@ -1,10 +1,12 @@
 {
 	"files": {
 		"includes": [
+			"src/**/*.cts",
 			"src/**/*.ts",
 			"test/**/*.ts",
-			"index.js",
 			"package.json",
+			"tsconfig.base.json",
+			"tsconfig.cjs.json",
 			"tsconfig.json",
 			"vitest.config.ts",
 			"biome.json"
@@ -12,7 +14,8 @@
 	},
 	"formatter": {
 		"enabled": true,
-		"indentStyle": "space"
+		"indentStyle": "space",
+		"indentWidth": 2
 	},
 	"linter": {
 		"enabled": true,
