+++ package.json
@@ -21,9 +21,9 @@
 		"library"
 	],
 	"scripts": {
-		"build": "tsx ./scripts/build.ts",
-		"lint": "tsx ./scripts/biome.ts check",
-		"format": "tsx ./scripts/biome.ts format --write",
+		"build": "rimraf dist && tsc -p tsconfig.json && tsc -p tsconfig.cjs.json",
+		"lint": "biome check .",
+		"format": "biome format --write .",
 		"prepublishOnly": "npm run build",
 		"test": "vitest run",
 		"test:unit": "vitest run --project unit",
@@ -36,10 +36,9 @@
 	},
 	"devDependencies": {
 		"@types/node": "^24.10.1",
-		"@biomejs/js-api": "^4.0.0",
-		"@biomejs/wasm-nodejs": "^2.4.13",
+		"@biomejs/biome": "^2.4.13",
 		"@vitest/coverage-v8": "^4.1.5",
-		"tsx": "^4.20.6",
+		"rimraf": "^6.0.1",
 		"typescript": "^6.0.3",
 		"vitest": "^4.1.5"
 	},
