+++ tsconfig.base.json
@@ -5,7 +5,6 @@
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"skipLibCheck": true,
-		"ignoreDeprecations": "6.0",
 		"types": ["node"],
 		"rootDir": "src"
 	},
